# lib/api-zod/src/generated/types/accommodationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AccommodationInputRoomType } from './accommodationInputRoomType';

export interface AccommodationInput {
  registrationId: number;
  roomType: AccommodationInputRoomType;
  /**
     * @minimum 1
     * @maximum 30
     */
  nights: number;
  notes?: string;
}