# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Registration } from './registration';

export interface DashboardSummary {
  participantName: string;
  nextRetreat: Registration | null;
  totalRetreats: number;
  prayerCount: number;
  certificateCount: number;
  notifications: number;
}