# lib/api-zod/src/generated/types/donationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DonationInput {
  donorName: string;
  email: string;
  /** @minimum 1 */
  amount: number;
  purpose: string;
}