# lib/api-zod/src/generated/types/prayerRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PrayerRequestStatus } from './prayerRequestStatus';

export interface PrayerRequest {
  id: number;
  fullName: string;
  email: string;
  request: string;
  isAnonymous: boolean;
  status: PrayerRequestStatus;
  createdAt: Date;
}