# lib/api-zod/src/generated/types/prayerRequestInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PrayerRequestInput {
  fullName: string;
  email: string;
  /** @minLength 10 */
  request: string;
  isAnonymous: boolean;
}