# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  retreatId: number;
  fullName: string;
  email: string;
  phone: string;
  /**
     * @minimum 1
     * @maximum 10
     */
  guests: number;
}