# lib/api-zod/src/generated/types/retreat.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RetreatStatus } from './retreatStatus';

export interface Retreat {
  id: number;
  title: string;
  slug: string;
  description: string;
  startDate: Date;
  endDate: Date;
  location: string;
  capacity: number;
  enrolled: number;
  price: number;
  status: RetreatStatus;
  spiritualTheme: string;
  imageUrl: string;
}