# lib/api-zod/src/generated/types/retreatUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RetreatInput } from './retreatInput';

export type RetreatUpdate = RetreatInput;